requirements.txt
ImageHash==4.3.1
pillow==10.2.0
python-dotenv==1.0.1